figma-backend/requirements.txt
fastapi==0.79.0
uvicorn==0.18.2
requests~=2.28.1
pydantic~=1.9.1
openai==0.22.1
gunicorn==20.1.0
pyyaml==6.0
